+++ src/Cipher.cpp
@@ -44,14 +44,35 @@ struct Cipher : Module {
 		NUM_LIGHTS
 	};
 
-	Cipher() {
+	Cipher()
+    {
 		config(NUM_PARAMS, NUM_INPUTS, NUM_OUTPUTS, NUM_LIGHTS);
+
+        configInput(CLOCK_INPUT, "Clock");
+        configInput(DATA1_INPUT, "Data 1");
+        configInput(DATA2_INPUT, "Data 2");
+        configInput(STROBE_INPUT, "Strobe");
+
+        configOutput(SERIAL_OUTPUT, "Super Serial");
+        configOutput(CV1_OUTPUT, "CV 1");
+        configOutput(CV2_OUTPUT, "CV 2");
+        configOutput(CV3_OUTPUT, "CV 3");
+        configOutput(CV4_OUTPUT, "CV 4");
+
+        configOutput(OOO_OUTPUT, "OOO Output");
+        configOutput(OOX_OUTPUT, "OOX Output");
+        configOutput(OXO_OUTPUT, "OXO Output");
+        configOutput(OXX_OUTPUT, "OXX Output");
+        configOutput(XOO_OUTPUT, "XOO Output");
+        configOutput(XOX_OUTPUT, "XOX Output");
+        configOutput(XXO_OUTPUT, "XXO Output");
+        configOutput(XXX_OUTPUT, "XXX Output");
 	}
 
 	bool bits[8] = {true, false, false, false, false, false, false, false};
     bool bitsOut[8] = {true, false, false, false, false, false, false, false};
     bool serialBit = false;
-    float cv1Value, cv2Value, cv3Value, cv4Value;
+    float cv1Value = 0.0f, cv2Value = 0.0f, cv3Value = 0.0f, cv4Value = 0.0f;
 
 	NLCTrigger clockIn;
 
@@ -219,11 +240,11 @@ struct CipherWidget : ModuleWidget {
         const float inJacksX = 7.0f;
         const float outJacksX = 56.75f;
         const float bottomJacksY = 106.0f;
-        int spacing = 10.0f;
+        float spacing = 10.0f;
 
         for (int i = 0; i < 4; i++)
         {
-            int xOffset = i * spacing;
+            float xOffset = i * spacing;
             addInput(createInput<PJ301MPort>(mm2px(Vec(inJacksX + xOffset, bottomJacksY)), module, Cipher::CLOCK_INPUT + i));
             addOutput(createOutput<PJ301MPort>(mm2px(Vec(outJacksX + xOffset, bottomJacksY)), module, Cipher::CV1_OUTPUT + i));
         }