+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+### 2.2.0
+- Add Sloth Torpor.
+- Add Sloth Apathy.
+- Add Sloth Intertia.
+- Add Triple Sloth.
+Many thanks to Don Cross for these excellent analog models!
+
+### 2.1.1
+- Fix Router SVG not being found on Linux.
+
+### 2.1.0
+- Add Router
+
+### 2.0.1
+- CMake build system support for developers.
+- Add Apple Silicon support to build script.
+- Add input and output port labels.
+- Fix phase of Divide and Conquer even clocks.
+
 ### 2.0.0
 - Build compatible with Rack 2.
 - Fixed missing param labels on Segue.